Create source files for each function
